+++ src/renderer/helpers/index.ts
@@ -3,6 +3,7 @@ import { useConstants } from './constants';
 import { useErrorHandling } from './errorHandling';
 import { useFormatters } from './formatters';
 import { useCurrentProject } from '../store/currentProject';
+import { useAppUpdate } from '../store/appUpdate';
 import { useRunProcess } from './runProcess';
 import { useUtilities } from './utilities';
 
@@ -14,8 +15,9 @@ export function useHelpers() {
 	const currentProject = useCurrentProject();
 	const runProcess = useRunProcess();
 	const utilities = useUtilities();
+	const appUpdate = useAppUpdate();
 
 	return {
-		api, constants, errors, formatters, currentProject, runProcess, utilities
+		api, constants, errors, formatters, currentProject, runProcess, utilities, appUpdate
 	}
 }