+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Revision 3.0.3 ###
 
+* Update tiledrain.str dist parameter default to 5m; users should manually update their project values if using tile drains
 * Bug fix in hard calibration and other pages where it wasn't allowing negative numbers
 * Bug fix in adding outflow to point source / inlets
 * Minor package updates